% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_matrices.R
\name{add_matrices}
\alias{add_matrices}
\title{Matrix addition}
\usage{
add_matrices(L)
}
\arguments{
\item{L}{A list of matrices.}
}
\value{
The sum of all matrices.
}
\description{
This function calculates the sum of several matrices.
}
\author{
Raphael Scherrer
}
