% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_spaces.R
\name{add_spaces}
\alias{add_spaces}
\title{Add spaces between words in labels}
\usage{
add_spaces(inames)
}
\arguments{
\item{inames}{A vector of character strings; the names to apply the function to.}
}
\value{
A vector of character strings, with spaces.
}
\description{
This function adds spaces between words to create clean labels for plotting.
}
\author{
Raphael Scherrer
}
