% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expvar.R
\name{calc_expvar}
\alias{calc_expvar}
\title{Explained variance of a principal component}
\usage{
calc_expvar(pca, n)
}
\arguments{
\item{pca}{A \code{prcomp} object.}

\item{n}{An integer: what principal component?}
}
\value{
A numeric, the proportion of variance in the dataset explained by the principal component.
}
\description{
This function calculates the explained variance of a given principal component in a PCA.
}
\author{
Raphael Scherrer
}
