% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_multinorm.R
\name{check_multinorm}
\alias{check_multinorm}
\title{Test multivariate normality across groups}
\usage{
check_multinorm(specdata, vars, plotit = T, method = "bonferroni")
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{plotit}{Logical. Whether to plot adjusted P-values.}

\item{method}{A character, the method used for P-value correction. See \code{?p.adjust}.}
}
\value{
A data frame with each group, its multivariate normality W statistic, P-value and adjusted P-value.
}
\description{
This function tests for multivariate normality of the data across groups (island-habitat combinations). If performs multiple Shapiro-Wilk's multivariate normality tests. P-values are adjusted to account for multiple testing.
}
\author{
Raphael Scherrer
}
