% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Test normality across groups}
\usage{
check_normality(specdata, vars, plotit = T, method = "bonferroni")
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{plotit}{Logical. Whether to plot adjusted P-values.}

\item{method}{A character, the method used for P-value correction. See \code{?p.adjust}.}
}
\value{
A list with one element per dependent variable. Each element is a data frame with Shapiro's W statistic, its P-value and adjusted P-value for each group (island-habitat combination). If \code{plotit = T}, also returns a bar plot of the adjusted P-values with a dashed line at 0.05.
}
\description{
This function tests for normality of the data across groups (island-habitat combinations). If performs multiple Shapiro-Wilk's normality tests. P-values are adjusted to account for multiple testing.
}
\author{
Raphael Scherrer
}
