% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatialcorr.R
\name{check_spatialcorr}
\alias{check_spatialcorr}
\title{Test spatial autocorrelation of the residuals}
\usage{
check_spatialcorr(specdata, vars, plotit = T, test = "moran",
  method = "bonferroni")
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{plotit}{Logical. Whether to plot residuals versus latitude and longitude.}

\item{test}{A character, what test to perform. \code{"moran"} (default) performs Moran's I test on each dependent variable. \code{"mantel"} performs Mantel's test to compare distance matrices in residual space and geographic space, meaning that it can perform a single test on multivariate residuals if several dependent variables are input.}

\item{method}{A character, the method used for P-value correction. See \code{?p.adjust}.}
}
\value{
Either the output of the \code{mantel.test} function from ape, or a table summarizing the (multiple) Moran's I test(s). If \code{plotit = T}, also returns a scatter plot of the residuals versus latitude and longitude.
}
\description{
This function tests for positive spatial autocorrelation of the (uni- or multivariate) residuals, using either Moran's I or Mantel's test.
}
\note{
Beware that Mantel's test is a permutation test (999 permutations), and will take longer tha Moran's I test to run.
}
\author{
Raphael Scherrer
}
