% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variances.R
\name{check_variances}
\alias{check_variances}
\title{Test homoscedasticity of the residuals across groups}
\usage{
check_variances(specdata, vars, method = "bonferroni")
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{method}{A character, the method used for P-value correction. See \code{?p.adjust}.}
}
\value{
A data frame with Bartlett's K^2 statistic, the degrees of freedom of the approximate chi-squares distribution, P-value and adjusted P-value, for each dependent variable.
}
\description{
This function tests for the homogeneity of the residuals across groups. It performs one Bartlett's test for each dependent variable. P-values are corrected.
}
\author{
Raphael Scherrer
}
