% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate_confusion_matrices.R
\name{cumulate_confusion_matrices}
\alias{cumulate_confusion_matrices}
\title{Cumulative confusion matrices}
\usage{
cumulate_confusion_matrices(svm.res)
}
\arguments{
\item{svm.res}{The output of \code{save_svm_results}.}
}
\value{
A list of cumulative SVM confusion matrices
}
\description{
Function to compute the cumulative confusion matrices of the Support Vector Machine analysis
}
\author{
Raphael Scherrer
}
