% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_anova.R
\name{dewlap_anova}
\alias{dewlap_anova}
\title{Multiple analyses of variance (ANOVA)}
\usage{
dewlap_anova(specdata, vars)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}
}
\value{
Nothing is returned. The output of the \code{anova} function is printed to the command prompt.
}
\description{
This function performs multiple two-way ANOVAs, one on each dependent variable. The effects of island, habitat and their interaction are tested.
}
\author{
Raphael Scherrer
}
