% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_dfa_global.R
\name{dewlap_dfa_global}
\alias{dewlap_dfa_global}
\title{Global Discriminant Function Analysis (DFA) of dewlap color}
\usage{
dewlap_dfa_global(specdata, vars, type = "linear", plotit = T,
  CV = F)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{type}{A character. Type of discriminant analysis. \code{"linear"} (default) or \code{"quadratic"}.}

\item{plotit}{Logical. Whether to plot the loadings of the data points on the discriminant functions (applicable only if \code{type == "linear"}).}

\item{CV}{Logical. Whether to jacknife the predictions (cross-validation).}
}
\value{
A vector with elements:
\itemize{
\item \code{observed}: the observed number of succesful assignments
\item \code{expected}: the number of succesful assignments expected by change
\item \code{n}: the number of assignments
\item \code{p.binom}: the P-value of the binomial test
\item \code{df}: the degrees of freedom of the MANOVA
\item \code{wilks}: Wilk's lambda
\item \code{approx.F}: the approximate F-value calculated from Wilk's lambda
\item \code{num.df}: numerator degrees of freedom for F-test
\item \code{denom.df}: denumerator degrees of freedom for F-test
\item \code{p.manova}: P-value of the MANOVA
}
}
\description{
This function performs a single, global DFA to test clustering by habitat across all islands. The algorithm fits functions that best discriminate among habitats based on dewlap color data. The type of DFA can be linear (LDA) or quadratic (QDA). Each data point is then classified i.e. its original habitat is predicted based on the discriminant functions. Predictions can be jacknifed (leave-one-out). The significance of the classification is assessed in two ways. First, a MANOVA tests for differences in dependent variables across predicted groups. Second, a binomial test assesses the departure of the observed number of successful predictions from the null expectation.
}
\note{
Quadratic discriminant analysis doesn't require homogeneous covariance matrices among groups, unlike linear (Robert I. Kabacoff, Quick-R, https://www.statmethods.net/advstats/discriminant.html).
}
\author{
Raphael Scherrer
}
