% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_gls.R
\name{dewlap_gls}
\alias{dewlap_gls}
\title{Fit and analyze generalized least squares models}
\usage{
dewlap_gls(specdata, vars, method = "bonferroni", plotit = F)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{method}{A character, the method used for P-value correction. See \code{?p.adjust}.}

\item{plotit}{Logical. Whether to plot the residuals of the best model or not.}
}
\value{
A list one element per dependent variable. Each element contains an AIC table comparing variance structures, and the ANOVA table summarizing the sequential LRTs of fixed effects. The function also plots the residuals (if \code{plotit = T}) and prints to the prompt the best model with the best variance structure.
}
\description{
This function fits GLS models to a set of dependent variables, one after the other. It compares several residual variance structures with AIC and retains the best. The fixed effects of the model are tested by model comparison using likelihood ratio tests.
}
\details{
GLS models are used to account for heteroscedasticity of the residuals across groups (Pinheiro and Bates 2000). The workflow here follows recommendations of Zuur et al. 2009.
}
\author{
Raphael Scherrer
}
