% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_permanova.R
\name{dewlap_permanova}
\alias{dewlap_permanova}
\title{Permutational multivariate analysis of variance (PERMANOVA) of dewlap color}
\usage{
dewlap_permanova(specdata, vars, nperm = 999)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{nperm}{Integer. The number of permutations to be computed.}
}
\value{
The output of the \code{adonis} function from vegan, containing the PERMANOVA results.
}
\description{
This function performs a two-way PERMANOVA of dewlap color data cross islands and habitats. It is the non-parametric version of MANOVA.
}
\author{
Raphael Scherrer
}
