% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2rspec.R
\name{df2rspec}
\alias{df2rspec}
\title{Convert to \code{rspec} with transposition}
\usage{
df2rspec(df)
}
\arguments{
\item{df}{A data frame of reflectance to convert. Columns for wavelengths must have "wl" preceding the wavelength in their name, e.g. "wl500" for 500nm. There is one row per measurement.}
}
\value{
An \code{rspec} data frame.
}
\description{
This function converts a reflectance data frame with wavelengths in columns into an \code{rspec} data frame (package \code{pavo}) with wavelengths as rows.
}
\author{
Raphael Scherrer
}
