% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Install dependencies of the package}
\usage{
install_dependencies(update = F)
}
\arguments{
\item{update}{If \code{TRUE} all packages in the list of dependencies are reinstalled. If not (default) only those that are missing.}
}
\description{
This function installs or update the dependencies of the sagreicolor package.
}
\author{
Raphael Scherrer
}
