% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.r
\name{linearize}
\alias{linearize}
\title{Smooth reflectance artifacts}
\usage{
linearize(specs, where, cols = 2:ncol(specs))
}
\arguments{
\item{specs}{An \code{rspec} data frame (pavo).}

\item{where}{A numeric vector of two numbers defining between what and what wavelength is the smoothing to be applied (e.g. \code{where = c(620, 630)} for smoothing between 620 and 630nm).}

\item{cols}{A vector of column numbers, indicating what measurements to apply the smoothing to. Defaults to all. Note: first measurement is column 2.}
}
\value{
An \code{rspec} data frame with artifacts removed.
}
\description{
This function removes artifactual peaks in reflectance profiles by applying linear smoothing on a portion of the profile.
}
\author{
Raphael Scherrer
}
