% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dataset.R
\name{make_dataset}
\alias{make_dataset}
\title{Make spectral dataset}
\usage{
make_dataset(what_habitats, spot = 2)
}
\arguments{
\item{what_habitats}{A character vector. What habitats to retain?}

\item{spot}{What sport on the dewlap? 1 = near the throat, 2 = center.}
}
\value{
A data frame, a subset of \code{specdata}.
}
\description{
This function produces the dataset from raw data.
}
\author{
Raphael Scherrer
}
