% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_color_space.R
\name{plot_color_space}
\alias{plot_color_space}
\title{Plot color in PC space}
\usage{
plot_color_space(specdata, plot_centroids = T)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "habitat".}

\item{plot_centroids}{Whether to plot only group centroids or all individual points.}
}
\description{
This wrap up function displays biplots along PC1, 2 and 3
}
\author{
Raphael Scherrer
}
