% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eigenvalues.R
\name{plot_eigenvalues}
\alias{plot_eigenvalues}
\title{PCA eigenvalues}
\usage{
plot_eigenvalues(specdata)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}
}
\value{
The cumulative distribution of explained variance.
}
\description{
This function plots the eigenvalues of the color space.
}
\author{
Raphael Scherrer
}
