% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot correlation between variables}
\usage{
plot_heatmap(specdata)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}
}
\value{
A correlation matrix.
}
\description{
This function plots a correlation heatmap between PCs and wavelengths
}
\author{
Raphael Scherrer
}
