% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neural_importance.R
\name{plot_neural_importance}
\alias{plot_neural_importance}
\title{Plot importance of variables in discriminating between habitats}
\usage{
plot_neural_importance(X, saveto, font)
}
\arguments{
\item{X}{The importance table returned by plot_neural.}

\item{saveto}{Where to save the PDF plot, optional}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\value{
Just plots.
}
\description{
This function plots the importance of non-wavelengths in discriminating habitats by the 5\% best neural networks
}
\author{
Raphael Scherrer
}
