% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neural_pvalues.R
\name{plot_neural_pvalues}
\alias{plot_neural_pvalues}
\title{Plot neural network classification}
\usage{
plot_neural_pvalues(X, saveto, font)
}
\arguments{
\item{X}{A data frame containing P-values for randomized and empirical data.}

\item{saveto}{Where to save the PDF plot, optional}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\value{
Just plots.
}
\description{
This function plots the P-values assessing the classification output of the neural networks.
}
\author{
Raphael Scherrer
}
