% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neural_spectrum.R
\name{plot_neural_spectrum}
\alias{plot_neural_spectrum}
\title{Plot importance of wavelengths in discriminating between habitats}
\usage{
plot_neural_spectrum(X, saveto, font)
}
\arguments{
\item{X}{The importance table returned by plot_neural.}

\item{saveto}{Where to save the PDF plot, optional}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\value{
Just plots.
}
\description{
This function plots the importance of wavelengths in discriminating habitats by the 5\% best neural networks
}
\author{
Raphael Scherrer
}
