% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neural_success.R
\name{plot_neural_success}
\alias{plot_neural_success}
\title{Plot neural network classification}
\usage{
plot_neural_success(X, saveto, font)
}
\arguments{
\item{X}{A data frame containing proportions of success for randomized and empirical data.}

\item{saveto}{Where to save the PDF plot, optional}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\value{
Just plots.
}
\description{
This function plots the classification performance of the neural networks.
}
\author{
Raphael Scherrer
}
