% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pc1_brightness.R
\name{plot_pc1_brightness}
\alias{plot_pc1_brightness}
\title{Plot PC1 against mean reflectance}
\usage{
plot_pc1_brightness(specdata)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}
}
\value{
The results of a Pearson's correlation test between the two.
}
\description{
This function plots PC1 against mean reflectance.
}
\author{
Raphael Scherrer
}
