% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot residuals of dewlap color}
\usage{
plot_residuals(specdata, vars)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}
}
\value{
Nothing. Just one (or more) scatter plot(s).
}
\description{
This function plots the residuals of a set of dependent variables across islands and habitats. Residuals are taken as the deviation from the group mean. Residuals are plotted against group means (fitted values).
}
\author{
Raphael Scherrer
}
