% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_spectra.R
\name{process_spectra}
\alias{process_spectra}
\title{Process reflectance spectra}
\usage{
process_spectra(specs, locale = NULL, spot = 1)
}
\arguments{
\item{specs}{An \code{rspec} data frame containing the reflectance profiles of all dewlaps.}

\item{locale}{A data frame containing locality information for each lizard (island, habitat, plot, longitude and latitude). Optional.}

\item{spot}{An integer: what spot to analyze? 1 (default) or 2.}
}
\value{
A summary data frame with processed spectra. Wavelengths are in columns, as well as extra details like island and habitats.
}
\description{
This function performs a few steps of processing on the raw reflectance data. These include removing juveniles and females, removing lizards from New Providence, keeping only one dewlap spot of interest, fixing reflectance artifacts, smoothing curves and removing outliers.
}
\author{
Raphael Scherrer
}
