% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_svm_results.R
\name{save_svm_results}
\alias{save_svm_results}
\title{Save SVM output}
\usage{
save_svm_results(neural.res, nested.res, saveto = ".")
}
\arguments{
\item{neural.res}{The output of \code{dewlap_neural}.}

\item{nested.res}{The output of \code{dewlap_neural_nested}.}

\item{saveto}{Where to save the output?}
}
\value{
A list of SVM outputs.
}
\description{
Function to save the results of the Support Vector Machine analysis
}
\author{
Raphael Scherrer
}
