% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_locations.R
\name{show_locations}
\alias{show_locations}
\title{Show geographical coordinates}
\usage{
show_locations()
}
\value{
A table with sampling locations.
}
\description{
Function to show exact sampling locations
}
\author{
Raphael Scherrer
}
