% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_svm_params.R
\name{show_svm_params}
\alias{show_svm_params}
\title{Show SVM parameters}
\usage{
show_svm_params(svm.res)
}
\arguments{
\item{svm.res}{The output of \code{save_svm_results}.}
}
\value{
A table of SVM parameters.
}
\description{
Function to display the parameters of the Support Vector Machine analysis
}
\author{
Raphael Scherrer
}
