% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_pca.R
\name{spectral_pca}
\alias{spectral_pca}
\title{Principal Component Analysis of dewlap reflectance}
\usage{
spectral_pca(specdata)
}
\arguments{
\item{specdata}{A data frame with reflectance in columns. The columns with reflectance information must have "wl" in their name, e.g. "wl500" for 500nm.}
}
\value{
A \code{prcomp} object.
}
\description{
This function calculates the principal components of a reflectance dataset.
}
\author{
Raphael Scherrer
}
