% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_habitats.R
\name{subset_habitats}
\alias{subset_habitats}
\title{Subset of the spectral dataset}
\usage{
subset_habitats(specdata, whatHabitats, inclusive = F)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{whatHabitats}{A character vector. What habitats to retain?}

\item{inclusive}{Do we include the islands with some but not all of the wanted habitats (Conception)? If \code{FALSE}, keep only the islands with all the wanted habitats.}
}
\value{
A data frame, a subset of \code{specdata}.
}
\description{
This function takes a subset of the spectral dataset in order to have no empty cell in the two-way analysis design. Given habitats to retain, the function eliminates all the observations that are not from these habitats, and eliminates all islands that do not have all these habitats present.
}
\author{
Raphael Scherrer
}
