% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multiContrasts.R
\name{test_multiContrasts}
\alias{test_multiContrasts}
\title{Test multivariate contrasts}
\usage{
test_multiContrasts(W, specdata, vars, method = "bonferroni")
}
\arguments{
\item{W}{A matrix indicating all contrasts to be tested, or a vector if only one contrast. The number of rows is the number of contrasts to test. The number of columns is the number of groups. The matrix/vector is filled with zeros, except for a 1 and a -1 at the position of the groups that are to be contrasted.}

\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{method}{Correction method for adjusting p-values.}
}
\value{
A data frame with the results of each Wilk's lambda test in rows. In columns,
\itemize{
\item{\code{Wilks} Wilk's lambda.}
\item{\code{approx.F} F-value computed from Wilk's lambda.}
\item{\code{df1}, \code{df2} Numerator and denominator degrees of freedom of the F distribution.}
\item{\code{p.value} P-value computed from the F-distribution.}
\item{\code{p.adj} Corrected P-value.}
}
}
\description{
This function tests for significant contrasts between multivariate habitat-means within islands. The procedure is parametric and involves calculation of Wilk's lambda and P-value adjustment.
}
\note{
The parametric procedure was adapted from Charles Zaiontz's post on multivariate contrast testing in Excel: http://www.real-statistics.com/multivariate-statistics/multivariate-analysis-of-variance-manova/manova-follow-up-contrasts/.
}
\author{
Raphael Scherrer
}
