% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_contrasts.R
\name{which_contrasts}
\alias{which_contrasts}
\title{Get significant contrasts}
\usage{
which_contrasts(pvalues, W, groups, alpha = 0.05)
}
\arguments{
\item{pvalues}{A vector of P-values from a multiple comparison procedure (see \code{?test_contrasts} and \code{?test_multiContrasts}).}

\item{W}{A matrix of contrast weights.}

\item{groups}{A vector of character strings. The names of the groups corresponding to the columns in \code{W} (the groups compared in the multiple comparison).}

\item{alpha}{A numeric. The significance threshold to declare a contrast significant. Defaults to 0.05.}
}
\value{
A table showing each pair of habitats that differed significantly in mean.
}
\description{
This function finds what habitat contrasts are significantly different from zero.
}
\author{
Raphael Scherrer
}
