% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Save output}
\usage{
write_output(results)
}
\arguments{
\item{results}{A list of results.}
}
\description{
Function to write output to text file
}
\author{
Raphael Scherrer
}
