% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_manova.R
\name{dewlap_manova}
\alias{dewlap_manova}
\title{Multivariate analysis of variance (MANOVA) of dewlap color}
\usage{
dewlap_manova(specdata, vars, perVariable = T)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{perVariable}{Logical. Whether to perform multiple two-way ANOVAs, one on each dependent variable. Defaults to \code{TRUE}.}
}
\value{
The MANOVA table is returned. The multiple ANOVAs, if applicable, are printed to the command prompt.
}
\description{
This function performs a two-way MANOVA of dewlap color data cross islands and habitats. It can also perform multiple two-way ANOVAs, one on each dependent variable (equivalent to the \code{dewlap_anova} function).
}
\author{
Raphael Scherrer
}
