% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_neural.R
\name{dewlap_neural}
\alias{dewlap_neural}
\title{Habitat classification with neural networks}
\usage{
dewlap_neural(specdata, vars, nRepet = 1000, saveto, seed, font)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{nRepet}{The number of neural networks to train (same number for empirical and permuted datasets).}

\item{saveto}{Where to save the PDFs. No PDF is saved if missing, plots will be returned in the R console.}

\item{seed}{Seed for random number gnerators}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\value{
Just plots.
}
\description{
This function trains neural networks to recognize differences between habitats. Each neural network is trained on a random sample of half of the data, and tested against the other half. The success of the classification is compared to a null expectation generated from a permuted dataset where no differences exist between habitats. The 5\% best performing machines are studied more in depth to identify what were the most important variables in discriminating between habitats.
}
\note{
If \code{plot_importance = T} and reflectances at certain wavelengths are present in the variables (their name start with "wl" in specdata) then a first linear plot is drawn along the light spectrum, then a barplot is generated with the remaining variables.
}
\author{
Raphael Scherrer
}
