% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_neural.R
\name{dewlap_neural}
\alias{dewlap_neural}
\title{Habitat classification with neural networks}
\usage{
dewlap_neural(specdata, vars, nRepet = 1000, plot_success = T,
  plot_importance = T, save_plot = T, seed)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{nRepet}{The number of neural networks to train (same number for empirical and permuted datasets).}

\item{plot_success}{Whether to plot histograms of the success of the machines on the empirical data compared to the null distribution.}

\item{plot_importance}{Whether to plot the importance of each variable in discriminating among habitats, according to the top 5\% machines.}

\item{save_plot}{Whether to save plots as pdfs to the plot/ folder or, if false, to display them in R.}

\item{seed}{Seed for random number gnerators}
}
\value{
Just plots.
}
\description{
This function trains neural networks to recognize differences between habitats. Each neural network is trained on a random sample of half of the data, and tested against the other half. The success of the classification is compared to a null expectation generated from a permuted dataset where no differences exist between habitats. The 5\% best performing machines are studied more in depth to identify what were the most important variables in discriminating between habitats.
}
\note{
If \code{plot_importance = T} and reflectances at certain wavelengths are present in the variables (their name start with "wl" in specdata) then a first linear plot is drawn along the light spectrum, then a barplot is generated with the remaining variables.
}
\author{
Raphael Scherrer
}
