% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewlap_neural_nested.R
\name{dewlap_neural_nested}
\alias{dewlap_neural_nested}
\title{Habitat classification with neural networks per island}
\usage{
dewlap_neural_nested(specdata, vars, nRepet = 1000, saveto, seed, font)
}
\arguments{
\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{nRepet}{The number of neural networks to train (same number for empirical and permuted datasets).}

\item{saveto}{Path to the folder where to save the PDFs. One folder will be created for each island.}

\item{seed}{Seed for random number generators}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\value{
Just plots.
}
\description{
This function applies the neural network analysis on each island separately
}
\author{
Raphael Scherrer
}
