% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_neural_importance.R
\name{plot_neural_importance}
\alias{plot_neural_importance}
\title{Plot importance of variables in discriminating between habitats}
\usage{
plot_neural_importance(importance, saveto, font)
}
\arguments{
\item{importance}{The importance table returned by \code{plot_neural} (the \code{imp} field).}

\item{saveto}{Where to save the PDF plot, optional}

\item{font}{Font to be used in the plots. Defaults to Helvetica.}
}
\description{
This function plots the importance of (wavelength and non-wavelength) variables in discriminating habitats by the 5\% best neural networks. Importance is defined as the cumulative weights of the input nodes.
}
\author{
Raphael Scherrer
}
