% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_contrasts.R
\name{test_contrasts}
\alias{test_contrasts}
\title{Test (multiple) univariate contrasts}
\usage{
test_contrasts(W, specdata, vars, parametric = T, plotit = T,
  method = "bonferroni")
}
\arguments{
\item{W}{A matrix indicating all contrasts to be tested, or a vector if only one contrast. The number of rows is the number of contrasts to test. The number of columns is the number of groups. The matrix/vector is filled with zeros, except for a 1 and a -1 at the position of the groups that are to be contrasted.}

\item{specdata}{A data frame containing at least columns for the dependent variables, as well as a column "island" and a column "habitat".}

\item{vars}{A character or integer vector. The names, or indices, of the dependent variables in \code{specdata}.}

\item{parametric}{Logical. If \code{TRUE} (default), parametric multiple comparisons are performed, otherwise a non-parametric procedure is used.}

\item{plotit}{Logical. If \code{TRUE}, plots 95% confidence intervals of the contrasts. One plot per dependent variable.}

\item{method}{Correction method for adjusting p-values (parametric procedure).}
}
\value{
A list with one element for each dependent variable in \code{vars}. Each element is either the output of the \code{summary.glht} function (parametric) or of the \code{nparcomp} function (non-parametric). It contains information about contrasts, P-values and confidence intervals. Also prints a table with all significant contrasts to the prompt.
}
\description{
This function tests for significant contrasts between habitat-means within islands in one or several dependent variables independently. The procedure is either parametric (package multcomp) or non-parametric (package nparcomp). The function can also plot the 95% confidence intervals of the differences in means.
}
\note{
Parametric workflow as per Salvatore S. Mangiafico (https://rcompanion.org/rcompanion/h_01.html). Non-parametric workflow from the package description of nparcomp. Note also that performing multiple comparisons one dependent variable at a time is not so bad if variables have limited correlation, as in the case of principal components (Krishnaiah and Reising 2006).
}
\author{
Raphael Scherrer
}
